new(demo): refactor line radial to codesandbox example
