+++ packages/vx-demo/src/pages/LineRadial.tsx
@@ -1,11 +1,15 @@
 import React from 'react';
 import Show from '../components/Show';
-import LineRadial from '../components/tiles/LineRadial';
-import LineRadialSource from '!!raw-loader!../components/tiles/LineRadial';
+import LineRadial from '../docs-v2/examples/vx-shape-line-radial/Example';
+import LineRadialSource from '!!raw-loader!../docs-v2/examples/vx-shape-line-radial/Example';
 
 export default () => {
   return (
-    <Show component={LineRadial} title="Line Radial">
+    <Show
+      component={LineRadial}
+      title="Line Radial"
+      codeSandboxDirectoryName="vx-shape-line-radial"
+    >
       {LineRadialSource}
     </Show>
   );
